fix: rename html example
