+++ apps/widget-configurator/src/app/embedDialog/utils/htmlExample.ts
